% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_argument_diff_length}
\alias{abort_argument_diff_length}
\title{Abort based on arguments having different lengths}
\usage{
abort_argument_diff_length(arg1, arg2)
}
\arguments{
\item{arg1, arg2}{A character vector with the argument name.}
}
\description{
Abort based on arguments having different lengths
}
\examples{
\dontrun{
x <- 1:5
y <- 1:10
if (lenght(x) != length(y)) {
  abort_argument_diff_length("x", "y")
}
}
}
\keyword{internal}
