% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_argument_length}
\alias{abort_argument_length}
\title{Abort based on the argument length}
\usage{
abort_argument_length(arg, must, not)
}
\arguments{
\item{arg}{A character string with the argument name.}

\item{must}{A character string specifying which length the arg must have.}

\item{not}{The argument name (unquoted). The function evaluates the length of
the argument and displays it in the error message.}
}
\description{
Abort based on the argument length
}
\examples{
\dontrun{
x <- 1:10
if (lenght(x) > 1) {
  abort_argument_length("x", must = "have length 1", not = x)
}
}
}
\keyword{internal}
