% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_argument_type}
\alias{abort_argument_type}
\title{Abort based on the argument type}
\usage{
abort_argument_type(arg, must, not)
}
\arguments{
\item{arg}{A character string with the argument name.}

\item{must}{A character string specifying which type the arg must have.}

\item{not}{The argument name (unquoted). The function evaluates the type of
the argument and displays it in the error message.}
}
\description{
Abort based on the argument type
}
\examples{
\dontrun{
x <- letters
if (!is.numeric(x)) {
  abort_argument_type("x", must = "be numeric", not = x)
}
}
}
\keyword{internal}
