% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_class_not_implemented}
\alias{abort_class_not_implemented}
\title{Abort method if class is not implemented}
\usage{
abort_class_not_implemented(fun, class)
}
\arguments{
\item{fun}{A character vector with the function name.}

\item{class}{A character vector with the class name.}
}
\description{
Returns a custom error condition created with
\code{\link[rlang:abort]{rlang::abort()}} with a - hopefully - more useful
error message and metadata.
}
\seealso{
\code{\link[=abort_argument_type]{abort_argument_type()}},
\code{\link[=abort_argument_length]{abort_argument_length()}},
\code{\link[=abort_argument_diff_length]{abort_argument_diff_length()}},
\code{\link[=abort_column_not_found]{abort_column_not_found()}}
}
