% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_column_not_found}
\alias{abort_column_not_found}
\title{Abort based on column not being found in a data frame}
\usage{
abort_column_not_found(data, col_name)
}
\arguments{
\item{data}{A data frame.}

\item{col_name}{A character vector with the column name.}
}
\description{
Abort based on column not being found in a data frame
}
\examples{
\dontrun{
data <- data.frame(x = 1:10)
if ("y" \%!in\% names(data)) {
  abort_column_not_found(data, "y")
}
}
}
\keyword{internal}
