% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_package_not_installed}
\alias{abort_package_not_installed}
\title{Abort if required package is not installed}
\usage{
abort_package_not_installed(package)
}
\arguments{
\item{package}{A character string with the required package name.}
}
\value{
Returns an error condition of class{error_package_not_installed}.
}
\description{
Creates a custom error condition created with
\code{\link[rlang:abort]{rlang::abort()}} with a - hopefully - more useful
error message and metadata.
}
\seealso{
\code{\link[=abort_argument_type]{abort_argument_type()}},
\code{\link[=abort_argument_class]{abort_argument_class()}},
\code{\link[=abort_argument_length]{abort_argument_length()}},
\code{\link[=abort_argument_diff_length]{abort_argument_diff_length()}},
\code{\link[=abort_column_not_found]{abort_column_not_found()}},
\code{\link[=abort_no_method_for_class]{abort_no_method_for_class()}}
}
