% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_means.R
\name{accuracy_means_params}
\alias{accuracy_means_params}
\title{Params for the accuracy indices summary functions}
\arguments{
\item{actual}{A numeric vector with the actual values.}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. Defaults to \code{FALSE}.}
}
\value{
Returns a double scalar with the accuracy index value.
}
\description{
Params for the accuracy indices summary functions
}
\keyword{internal}
