% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_means.R
\name{bias}
\alias{bias}
\title{Bias}
\usage{
bias(actual, predicted, na.rm = FALSE)
}
\arguments{
\item{actual}{A numeric vector with the actual values}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. Defaults to \code{FALSE}}
}
\description{
Computes the bias (mean error) between the input vectors.
}
\seealso{
\code{\link[=mean_error]{mean_error()}},
\code{\link[=loa]{loa()}}
}
