% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman_plot.R
\name{bland_altman_plot}
\alias{bland_altman_plot}
\title{Create a Bland-Altman plot}
\usage{
bland_altman_plot(x, ...)
}
\arguments{
\item{x}{An object of class \code{lvmisc_cv} or an object containing a model.}

\item{...}{Additional arguments to be passed to \code{ggplot2::aes()}.}
}
\description{
Create a Bland-Altman plot as described by
\href{=https://bit.ly/3mbf8bL}{Bland & Altman (1986)}.
}
\examples{
mtcars <- tibble::as_tibble(mtcars, rownames = "car")
m <- stats::lm(disp ~ mpg, mtcars)
cv <- loocv(m, mtcars, car)
bland_altman_plot(cv, colour = as.factor(am))
}
