% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Compute body mass index (BMI)}
\usage{
bmi(data, mass, height, category = FALSE)
}
\arguments{
\item{data}{The data frame or tibble with the data to be computed.}

\item{mass, height}{The bare (unquoted) names of the columns with body mass
and height data. \code{mass} unit must be kilograms and \code{height} unit must be
meters. If the \code{height} unit is centimeters, it is converted to meters
before BMI computation and a warning is shown.}

\item{category}{If TRUE, the \code{\link[=bmi_cat]{bmi_cat()}} function is called to
classify the BMI category, returning the classification in another column
as a factor; otherwise (the default) BMI category is not classified.}
}
\description{
\code{bmi} returns the row-wise BMI in kilograms per meter squared.
}
\examples{
df <- dplyr::starwars \%>\% 
  # Converting height to meters
  dplyr::mutate(height = height / 100)

bmi(df, mass, height, category = TRUE)
}
\seealso{
\code{\link[=bmi_cat]{bmi_cat()}}
}
