% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{check_package}
\alias{check_package}
\title{Checks whether a package is installed}
\usage{
check_package(x)
}
\arguments{
\item{x}{A character string with the package name}
}
\value{
If all packages in \code{x} are installed, returns \code{TRUE},
if not, returns the name of the non-installed package(s).
}
\description{
Checks whether a package is installed
}
\keyword{internal}
