% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{error}
\alias{error}
\title{Error}
\usage{
error(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector with the actual values}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}
}
\description{
Computes the element-wise error between the input vectors.
}
\examples{
actual <- runif(10)
predicted <- runif(10)

error(actual, predicted)
}
\seealso{
\code{\link[=error_pct]{error_pct()}},
\code{\link[=error_abs]{error_abs()}},
\code{\link[=error_abs_pct]{error_abs_pct()}},
\code{\link[=error_sqr]{error_sqr()}}.
}
