% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{error_abs_pct}
\alias{error_abs_pct}
\title{Absolute percent error}
\usage{
error_abs_pct(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector with the actual values}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}
}
\value{
A vector of the class \code{lvmisc_percent} with the element-wise
absolute percent error values.
}
\description{
Computes the element-wise absolute percent errors between the input vectors.
}
\examples{
actual <- runif(10)
predicted <- runif(10)

error_abs_pct(actual, predicted)
}
\seealso{
\code{\link[=error]{error()}},
\code{\link[=error_pct]{error_pct()}},
\code{\link[=error_abs]{error_abs()}},
\code{\link[=error_sqr]{error_sqr()}}.
}
