% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{error_pct}
\alias{error_pct}
\title{Percent error}
\usage{
error_pct(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector with the actual values}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}
}
\value{
Returns a double vector with the element-wise percent error values.

A vector of the class \code{lvmisc_percent} with the element-wise
percent error values.
}
\description{
Computes the element-wise percent error between the input vectors.
}
\examples{
actual <- runif(10)
predicted <- runif(10)

error_pct(actual, predicted)
}
\seealso{
\code{\link[=error]{error()}},
\code{\link[=error_abs]{error_abs()}},
\code{\link[=error_abs_pct]{error_abs_pct()}},
\code{\link[=error_sqr]{error_sqr()}}.
}
