% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv.R
\name{get_cv_fixed_eff}
\alias{get_cv_fixed_eff}
\alias{get_cv_r2}
\title{Extract information from the trained models from a cross-validation}
\usage{
get_cv_fixed_eff(cv)

get_cv_r2(cv)
}
\arguments{
\item{cv}{An object of class \code{lvmisc_cv}.}
}
\value{
\code{get_cv_fixed_eff()} returns a tibble with the estimated
value for each coefficient of each trained model and its associated
standard error. \code{get_cv_r2()} returns a tibble with the R squared
for each of the trained models.
}
\description{
Extract information from the trained models from a cross-validation
}
