% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{loocv}
\alias{loocv}
\alias{loocv.default}
\alias{loocv.lm}
\title{Leave-one-out cross-validation}
\usage{
loocv(model, data, id, keep = "used")

\method{loocv}{default}(model, data, id, keep = "used")

\method{loocv}{lm}(model, data, id, keep = "used")
}
\arguments{
\item{model}{A model object.}

\item{data}{A data frame.}

\item{id}{The bare (unquoted) name of the column which identifies subjects.}

\item{keep}{A character string indicating which variables to keep.}
}
\description{
Leave-one-out cross-validation
}
