% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lt}
\alias{lt}
\alias{le}
\title{Last error}
\usage{
lt()

le()
}
\description{
\code{lt()} prints the last error and the full backtrace and \code{le()}
returns the last error with a simplified backtrace. These functions are
just wrappers to \code{\link[rlang:last_trace]{rlang::last_trace()}} and
\code{\link[rlang:last_error]{rlang::last_error()}} respectively.
}
