% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lu}
\alias{lu}
\alias{lna}
\title{Number of elements in a vector.}
\usage{
lu(x)

lna(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\code{lu} returns the number of non-\code{NA} unique elements and \code{lna}
returns the number of \code{NA}s.
}
\examples{
x <- sample(c(1:3, NA), 10, replace = TRUE)
lu(x)
lna(x)
}
\seealso{
\code{\link[base:length]{length()}},
\code{\link[base:unique]{unique()}},
\code{\link[base:is.na]{is.na()}}
}
