% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_means.R
\name{mean_error_abs_pct}
\alias{mean_error_abs_pct}
\title{Mean absolute percent error}
\usage{
mean_error_abs_pct(actual, predicted, na.rm = TRUE)
}
\arguments{
\item{actual}{A numeric vector with the actual values}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
A vector of the class \code{lvmisc_percent}.
}
\description{
Computes the average absolute percent error between the input vectors.
}
\seealso{
\code{\link[=mean_error]{mean_error()}},
\code{\link[=mean_error_abs]{mean_error_abs()}},
\code{\link[=mean_error_pct]{mean_error_pct()}},
\code{\link[=mean_error_sqr]{mean_error_sqr()}},
\code{\link[=mean_error_sqr_root]{mean_error_sqr_root()}}
}
