% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_method.R
\name{new_lvmisc_accuracy}
\alias{new_lvmisc_accuracy}
\title{Constructor for lvmisc_accuracy object}
\usage{
new_lvmisc_accuracy(accuracy_data, model_class)
}
\arguments{
\item{accuracy_data}{A data frame with accuracy indices.}

\item{model_class}{The class of the model.}
}
\description{
Constructor for lvmisc_accuracy object
}
\keyword{internal}
