% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pa}
\alias{pa}
\title{Print all rows of a data frame or tibble}
\usage{
pa(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
Prints \code{data} and returns it invisibly.
}
\description{
Shortcut to print all rows of a data frame or tibble. Useful to inspect the
whole tibble, as it prints by default only the first 20 rows.
}
\examples{
df <- dplyr::starwars
pa(df)
}
\seealso{
\code{\link[base:print]{print()}} and
\href{https://tibble.tidyverse.org/reference/formatting.html}{printing tibbles}.
}
