% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_class.R
\name{percent}
\alias{percent}
\alias{is_percent}
\alias{as_percent}
\title{\code{percent} vector}
\usage{
percent(x = double())

is_percent(x)

as_percent(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{percent()}: A numeric vector
\item For \code{is_percent()}: An object to test.
\item For \code{as_percent()}: An object to cast.
}}
}
\value{
An S3 vector of class \code{lvmisc_percent}.
}
\description{
Creates a double vector that represents percentages. When printed, it is
multiplied by 100 and suffixed with \verb{\%}.
}
\examples{
percent(c(0.25, 0.5, 0.75))
}
