% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{Computes the percent change}
\usage{
percent_change(data, baseline, followup)
}
\arguments{
\item{data}{The data frame or tibble with the data to be computed.}

\item{baseline, followup}{The bare (unquoted) name of the columns to be
used to compute the percent change.}
}
\description{
\code{percent_change} returns the row-wise percent change between two
columns in a data frame.
}
\examples{
df <- data.frame(a = sample(20:40, 10))
df$b <- df$a * runif(10, min = 0.5, max = 1.5) 

percent_change(df, a, b)
}
