% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort_column_not_found}
\alias{abort_column_not_found}
\title{Abort based on column not being found in a data frame}
\usage{
abort_column_not_found(data, col_name)
}
\arguments{
\item{data}{A data frame.}

\item{col_name}{A character vector with the column name.}
}
\value{
Returns an error condition of class{error_column_not_found}.
}
\description{
Creates a custom error condition created with
\code{\link[rlang:abort]{rlang::abort()}} with a - hopefully - more useful
error message and metadata.
}
\seealso{
\code{\link[=abort_argument_type]{abort_argument_type()}},
\code{\link[=abort_argument_class]{abort_argument_class()}},
\code{\link[=abort_argument_length]{abort_argument_length()}},
\code{\link[=abort_argument_diff_length]{abort_argument_diff_length()}},
\code{\link[=abort_no_method_for_class]{abort_no_method_for_class()}},
\code{\link[=abort_package_not_installed]{abort_package_not_installed()}}
}
