% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_loocv.R
\name{accuracy}
\alias{accuracy}
\alias{accuracy.default}
\alias{accuracy.loocv}
\alias{accuracy.lm}
\title{Model accuracy}
\usage{
accuracy(x, na.rm = FALSE)

\method{accuracy}{default}(x, na.rm = FALSE)

\method{accuracy}{loocv}(x, na.rm = FALSE)

\method{accuracy}{lm}(x, na.rm = FALSE)
}
\description{
Model accuracy
}
