% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy_params}
\alias{accuracy_params}
\title{Params for the accuracy indices functions}
\arguments{
\item{actual}{A numeric vector with the actual values}

\item{predicted}{A numeric vector with the predicted values. Each element in
this vector must be a prediction for the corresponding element in
\code{actual}.}
}
\value{
Returns a double vector with the element-wise accuracy index values.
}
\description{
Params for the accuracy indices functions
}
\keyword{internal}
