% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi_cat}
\alias{bmi_cat}
\title{Classify body mass index (BMI) category}
\usage{
bmi_cat(data, bmi)
}
\arguments{
\item{data}{The data frame or tibble with the data to be computed.}

\item{bmi}{The bare (unquoted) name of the column with BMI data.}
}
\description{
\code{bmi_cat} returns the row-wise BMI category as factor with 6 levels:
\itemize{
\item Underweight (18.5 < BMI)
\item Normal weight (18.5 \eqn{\le} BMI < 25)
\item Overweight (25 \eqn{\le} BMI < 30)
\item Obesity class I (30 \eqn{\le} BMI < 35)
\item Obesity class II (35 \eqn{\le} BMI < 40)
\item Obesity class III (40 \eqn{\le} BMI)
}
}
\examples{
df <- dplyr::starwars \%>\% 
  # Converting height to meters
  dplyr::mutate(height = height / 100) \%>\% 
  bmi(mass, height)

bmi_cat(df, bmi)
}
\seealso{
\code{\link[=bmi]{bmi()}}
}
