% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv.R
\name{new_lvmisc_cv}
\alias{new_lvmisc_cv}
\title{Constructor for lvmisc_cv object}
\usage{
new_lvmisc_cv(x, model, trained_models)
}
\arguments{
\item{x}{A data.frame.}

\item{model}{A list of all trained models.}
}
\description{
Constructor for lvmisc_cv object
}
\keyword{internal}
