% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{new_lvmisc_cv}
\alias{new_lvmisc_cv}
\title{Constructor for loocv object}
\usage{
new_lvmisc_cv(x, model)
}
\arguments{
\item{x}{A data.frame.}

\item{model}{An object containing a model.}
}
\description{
Constructor for loocv object
}
\keyword{internal}
