% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\alias{percent_change.data.frame}
\title{Compute percent change between two vectors.}
\usage{
percent_change(baseline, followup)

\method{percent_change}{data.frame}(data, baseline, followup)
}
\arguments{
\item{baseline}{Name of the column with the baseline values}

\item{followup}{Name of the column with the follow-up values}

\item{data}{A data frame}
}
\value{
A numeric vector with the percent change.

The data.frame with a new column with percent change values
}
\description{
Compute percent change between two vectors.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: S3 method for data.frame
}}

