% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\alias{plot_model_multicollinearity}
\alias{plot_model_residuals}
\title{Plot model diagnostics}
\usage{
plot_model_multicollinearity(model)

plot_model_residuals(model)
}
\arguments{
\item{model}{An object containing a model.}
}
\description{
Plotting functions for some common model diagnostics.
}
\details{
\code{plot_model_multicollinearity} plots a bar chart of the
variance inflation factor (VIF) for each of the model terms.
\code{plot_model_residuals} plots a QQ plot of the model standardized
residuals.
}
\examples{
m <- lm(disp ~ mpg + hp + cyl + mpg:cyl, mtcars)
plot_model_multicollinearity(m)
plot_model_residuals(m)
}
