% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_multicollinearity.R
\name{plot_model}
\alias{plot_model}
\alias{plot_model_multicollinearity}
\title{Plot model diagnostics}
\usage{
plot_model_multicollinearity(model)
}
\arguments{
\item{model}{An object containing a model.}
}
\description{
\code{plot_model_multicollinearity} plots a bar chart of the variance
inflation factor (VIF) for each of the model terms.
}
\examples{
m <- lm(disp ~ mpg + hp + cyl + mpg:cyl, mtcars)
plot_model_multicollinearity(m)
}
