% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\alias{vif.default}
\alias{vif.lm}
\alias{vif.lmerMod}
\title{Compute variance inflation factor}
\usage{
vif(model)

\method{vif}{default}(model)

\method{vif}{lm}(model)

\method{vif}{lmerMod}(model)
}
\description{
Compute variance inflation factor
}
